==> Step.4_Feature correlation.ipynb <==
# Step.4 Feature Correlation:

**4.1 Import neccessary libraries** 

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

**4.2  Read the dataset** 

df = pd.read_csv("Feature.csv")
df.dtypes

#Exclude columns of type 'object'.
drop_colmns = ['payment_types', 'order_status', 'product_category_name']
df = df.drop(drop_colmns, axis=1)

#Convert the five-point scale to binary: 1 for high score group and 0 for low score group.
df['review_score'] = df['review_score'].apply(lambda x: 1 if x >= 4 else 0)

df.isnull().sum()

**4.3 Correlation between features and review scores, as well as inter-feature correlation** 

# Calculate the correlation matrix
corr_matrix = df.corr()
np.fill_diagonal(corr_matrix.values, np.nan)
# Get the correlation coefficients for the target variable 'review_score'
review_corr = corr_matrix['review_score']

**4.3.1 Positive correlations**

print("Variables with positive correlation:")
print(review_corr[review_corr > 0].sort_values(ascending=False))

**4.3.2 Negative correlations**

print("Variables with negative correlation:")
print(review_corr[review_corr < 0].sort_values())

**4.3.3 Select the top 5 variables with the strongest correlations (absolute values):**

top_features = review_corr.abs().sort_values(ascending=False).head(11)
print("Top variables with the strongest correlations (absolute values):")
print(top_features)

**4.3.4 Identify highly correlated features (correlation greater than 0.9 or less than -0.9).**

highly_correlated_features = corr_matrix[(corr_matrix > 0.9) | (corr_matrix < -0.9)].stack().index.tolist()
print("Highly correlated features:")
print(highly_correlated_features)
